Create Print On Screen The User Details
The code you provided defines an empty unordered list element (UL) in the HTML with an ID of "listOfitems":

bash
Copy code
<ul id="listOfitems"></ul>
The JavaScript function showUserOnScreen takes an object as its argument and displays its properties (name, email, and phone number) as a new list item (LI) in the unordered list element. Here is what the code inside the function does:

javascript
Copy code
const parentElem = document.getElementById('listOfitems');  // Get the unordered list element by its ID
parentElem.innerHTML = parentElem.innerHTML+ `<li>${obj.name} - ${obj.email} - ${obj.phonenumber}</li>`; // Add a new list item to the unordered list element
The first line of the function gets the unordered list element by its ID using the getElementById method from the document object. It stores the unordered list element in the parentElem variable.

The second line uses the innerHTML property of the unordered list element to add a new list item to the end of the existing list items. The new list item is created using a template literal, which allows the values of the object's name, email, and phonenumber properties to be inserted into the string. The ${} syntax is used to indicate the locations where these values should be inserted.

The innerHTML property of the unordered list element is updated with the new list item, which is then displayed on the screen.